`a`, and `b` should be `double` and not `int`

+++ src/Function.java
@@ -4,10 +4,10 @@ public abstract class Function{
     public abstract double valueAt(double point);
     public abstract String toString();
     public abstract Function derivative();
-    public abstract double bisectionMethod(int a, int b, double epsilon);
-    public abstract double bisectionMethod(int a, int b);
-    public abstract double newtonRaphsonMethod(int a, double epsilon);
-    public abstract double newtonRaphsonMethod(int a);
+    public abstract double bisectionMethod(double a, double b, double epsilon);
+    public abstract double bisectionMethod(double a, double b);
+    public abstract double newtonRaphsonMethod(double a, double epsilon);
+    public abstract double newtonRaphsonMethod(double a);
     public abstract String taylorPolynomial(int n);
 
 }
